requirements.txt
# HTTP Clients
requests==2.31.0
httpx==0.25.0

# Async support
asyncio==3.4.3
aiohttp==3.8.6
asyncpg==0.30.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23
alembic==1.12.1

# Cache
redis==5.0.1
hiredis==2.2.3

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
structlog==23.2.0
loguru==0.7.2

# Scheduling
APScheduler==3.10.4

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3